+++ services/avenirs-portfolio/scripts/avenirs-portfolio-bootstrap.sh
@@ -83,14 +83,10 @@ echo "spring.datasource.url=jdbc:postgresql://$AVENIRS_POSTGRESQL_PRIMARY_CONTAI
 [ "`hostname`" = "srv-dev-avenir" ] && swagger_root="srv-dev-avenir.srv-avenir.brgm.recia.net" || swagger_root="localhost"
 echo "app.server.url=http://$swagger_root/avenirs-portfolio-security" >> $AVENIRS_PORTFOLIO_SECURITY_SPRING_ENV_FILE;
 
-info "Database initialization files generation for avenirs-portfolio-security" 
-JASYPT_UTIL_SCRIPT=$AVENIRS_PORTFOLIO_SCRIPT_DIR/../avenirs-portfolio-security/scripts/jasypt-decrypt
-AVENIRS_PORTFOLIO_SECURITY_CLEAN_DB=$AVENIRS_PORTFOLIO_SCRIPT_DIR/../avenirs-portfolio-security/src/main/resources/db/clean.sql
-AVENIRS_PORTFOLIO_SECURITY_INIT_DB=$AVENIRS_PORTFOLIO_SCRIPT_DIR/../avenirs-portfolio-security/src/main/resources/db/init-db.sql
-AVENIRS_PORTFOLIO_SECURITY_CLEAN_TEST_DB=$AVENIRS_PORTFOLIO_SCRIPT_DIR/../avenirs-portfolio-security/src/test/resources/db/clean-test-db.sql
-AVENIRS_PORTFOLIO_SECURITY_INIT_TEST_DB=$AVENIRS_PORTFOLIO_SCRIPT_DIR/../avenirs-portfolio-security/src/test/resources/db/init-test-db.sql
 
 # Database initialization files generation
+info "Database initialization files generation for avenirs-portfolio-security" 
+
 cat $AVENIRS_PORTFOLIO_SECURITY_CLEAN_DB  > $AVENIRS_PORTFOLIO_SECURITY_CLEAN_DB_CLEAR
 vverbose "Created $AVENIRS_PORTFOLIO_SECURITY_CLEAN_DB_CLEAR"